+++ shodan/__init__.py
@@ -1,2 +1,6 @@
 from shodan.client import Shodan
 from shodan.exception import APIError
+
+__version__ = '1.32.0'
+
+__all__ = ['Shodan', 'APIError', '__version__']